upload implemented; updated  app style

+++ file_browser/index.php
@@ -1,54 +1,104 @@
+
 <!DOCTYPE html>
 <html lang="en">
 <head>
+<link rel="stylesheet" href="style.css">
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Document</title>
-    <style>
-table {
-    font-family: Arial, Helvetica, sans-serif;
-  border-collapse: collapse;
-  width: 100%;
-}
-table td, table th {
-  border: 1px solid #ddd;
-  padding: 8px;
-}
-table tr:nth-child(even){background-color: #f2f2f2;}
-table tr:hover {background-color: #ddd;}
-table th {
-  padding-top: 12px;
-  padding-bottom: 12px;
-  text-align: left;
-  background-color: #556B2F;
-  color: white;
-}
+    <title>File_browser</title>
+    <?php
+session_start();
+//logout
+(isset($_POST['logout']) && $_POST['logout'] == 'Logout')
+?  session_destroy() && session_start() 
+: null;
+//login user
+$note= '';
+(isset($_POST['login']) && $_POST['username'] !== "" && $_POST['password'] !=='') 
+? ((($_POST['username'] == "Guest" && $_POST['password'] == "iamaguest")) || ((($_POST['username'] == "Admin" && $_POST['password'] == "iamaadmin")))//loggin details
+? $_SESSION['logged_in'] = true  && $_SESSION['username'] = $_POST['username']
+: $note = 'Invalid username or password')
+: null;
 
-h5{
-    font-size: 24px;
-}
-span{
-    font-size: 24px;
-}
-    </style>
-
+?>
 </head>
 <body>
-<table>
-<tr>
-    <th>Type</th>
-    <th>Name</th>
-    <th>Actions</th>
-  </tr>
-  <?php
-  
-showFiles();
-  ?>
+    <div class="main">
+    <header>
+        <div class="container">
+        <?php
+      isset($_SESSION['logged_in']) and ($_SESSION['logged_in'] == true)
+     ? print  "<h3> Welcome, " . $_SESSION['username'] . "</h3>"
+     : null;
+      ?>
+    
+     <form class='logout' method='post' action='' <?php isset($_SESSION['logged_in']) && $_SESSION['logged_in'] == true
+         ? print("style = 'display: block'")
+         : print("style = 'display: none'") ?>
+         > 
+         <input type='submit' name='logout' value='Logout'>  
+        </form>  
+</div>
+<div class = "container2">
+        <form  action = "" method = "POST" enctype = "multipart/form-data">
+        <input type = "file"  id="upload" name = "image" />
+        <input type = "submit" class="greenBtn"/>
+    </form>
+    <?php 
+    if(isset($_FILES['image'])){
+        $errors= array();
+        $file_name = $_FILES['image']['name'];
+        $file_size = $_FILES['image']['size'];
+        $file_tmp = $_FILES['image']['tmp_name'];
+        $file_type = $_FILES['image']['type'];
+        if($file_size > 41943039) {
+            $errors[]='File size must be smaller than 40 MB'; 
+        }
+        if(empty($errors)==true) {
+            move_uploaded_file($file_tmp,  modifiedPath () . $file_name);
+            print "<p> Your file successfully uploaded!</p>";
+        }
+        else{
+            print_r($errors);
+        }
+    }
+ 
+    ?>
+    
+    </div><div class="dirLine" > <span><b> Currently in </b> <?php print modifiedPath () ?>  <b>directory</b></span> 
+</div>
+    </header>
 
+        <?php  
+             
+        if (isset($_SESSION['logged_in']) && $_SESSION['logged_in'] == true){
+           print  ("<section><table>
+            <tr>
+                <th>Type</th>
+                <th>Name</th>
+                <th>Actions</th>
+              </tr>");
+            showFiles();
+            print "</table>";
+            generateForm();
+            print "</section>";
+        } ?>
+    
+    <div class='logIn'>    
+        <form class='login' action="" method="post"
+        <?php isset($_SESSION['logged_in']) && $_SESSION['logged_in'] == true
+         ? print("style = 'display: none'")
+         : print("style = 'display: block'") ?>
+      >
+      <h4>You have to log in to see content</h4>
+      <h6><?php echo $note; ?></h6>
+        <input type="text" name="username" placeholder="Please enter username" required autofocus></br>
+        <input type="password" name="password" placeholder="Please enter password" required>
+        <button type="submit" name="login">Login</button>
+    </form>
+    </div>
 <?php
-echo '<span> <b> Currently in </b>' . $_SERVER['REQUEST_URI'] .' <b>dir</b> </span>';
-
 
 function modifiedPath (){
     $createFileInDir =  $_SERVER['DOCUMENT_ROOT'] . $_SERVER['REQUEST_URI'];
@@ -65,35 +115,40 @@ function showFiles() {
     $files_within = scandir($directory);
     $files_within = array_diff($files_within,array('.','..'));
     $files_within = array_values($files_within);
- 
-    foreach ($files_within as $key => $_file) {
+    
+    foreach ($files_within as $_file) {
         $uriLink = str_replace(" ", "%20", $_file);
         $pathLink = str_replace(" ", "%20", $path);
         $hrefPath = "<a href=?path=";
-        $deleteButton = "<form method='POST' action=''>
+        $deleteButton = "<td><form class='delete' method='POST' action=''>
         <input type='hidden' name='_fileDelete' value='$_file'>
         <input type='submit' name='delete' value='Delete'>
-        </form>";
+        </form></td>";
     
              is_dir($directory . $_file)
              ? (!isset($_GET["path"]) 
-             ? print "<tr><td>" . "Directory" . "</td>" . "<td>" . $hrefPath . $uriLink . "/>" . $_file . "</td><td>$deleteButton</td></tr>"
-             : print "<tr><td>Directory</td><td>" . $hrefPath . $pathLink . $uriLink . "/>" . $_file . "</td><td>$deleteButton</td></tr>")
-             : print ("<tr><td>File</td><td>" .  $_file . "</td><td>$deleteButton</td></tr>");
-        } 
-}
-?>
-</table>
-<form method="post" action=""> 
-<input type="hidden" name="back" value="Back">
-<input type="submit" name="submit" value="Go Back">  
-</form>
-<form method="post" action=""> 
-<input type="text" name="newFile" placeholder="Enter file/dir name" value="">
-<input type="submit" name="create" value="Submit">  
-</form>
+             ? print "<tr><td>Directory</td><td>" . $hrefPath . $uriLink . "/>" . $_file . "</td>$deleteButton</tr>"
+             : print "<tr><td>Directory</td><td>" . $hrefPath . $pathLink . $uriLink . "/>" . $_file . "</td>$deleteButton</tr>")
+             : print ("<tr><td>File</td><td>" .  $_file . "</td>$deleteButton</tr>");
+        }   
+    }
+function generateForm() {
+print "
+<div class='forms'>
+<div>
+<form class='back' method='post' action=''> 
+<input type='hidden' name='back' value='Back'>
+<input type='submit' name='submit' value='Go Back'>  
+</form></div><div>
+<form class='create'method='post' action=''> 
+<input id='field' type='text' name='newFile' placeholder='Enter file/dir name' value=''>
+<input type='submit' name='create' value='Create'>  
+</form></div></div>";
 
-<?php
+addDeleteBack();
+}
+function addDeleteBack() {
+
 if(isset($_POST['back'])){
     header("Location:" . (dirname($_SERVER['REQUEST_URI'])). '/' );
    }
@@ -104,7 +159,7 @@ if(isset($_POST['back'])){
         
         if (!file_exists($path . $_newFile )){
             if(is_dir($path)){
-                print $path . $_newFile;
+                // print $path . $_newFile;
                fopen($path . $_newFile, "w");
                header("Refresh:0");
             }
@@ -127,13 +182,11 @@ if(isset($_POST['back'])){
         }
        }
    }
-?>
-<?php
 if(isset($_POST['delete'])){
     $fileToDelete = $_POST['_fileDelete'];
     $path= modifiedPath ();
 
-    if($fileToDelete!=="index.php"){
+    if($fileToDelete!=="index.php" && $fileToDelete!=="style.css"){
         if (is_file($path . $fileToDelete)) {
             if (!unlink($path . $fileToDelete)) {
                 header("Refresh:0");
@@ -157,15 +210,14 @@ if(isset($_POST['delete'])){
 
     }
     else{
-        print  (" <p style='color:red; font-size:32px; font-weight:bold'> Are you crazy??? index.php is forbidden to delete!!! </p>");
+        print  (" <p style='color:red; font-size:32px; font-weight:bold'> Are you crazy??? Neither index.php nor style.css is forbidden to delete!!! </p>");
     }
-
 }
-
+}
 ?>
+</div>
 </body>
 </html>
-
